Introduced functionality for passcheck
This introduces the core-functionality for passcheck
and documents the behavior when certain things are missing
such as schema.

The overall idea is for the user to insert the function they
want to be associated with the passcheck feature. Then when
the hook is called, we will via SPI_exec read the table, and
call out to the function.

This requires SELECT privileges for all users on the bc.feature_info
table and EXECUTE permissions on the function inserted into the table.
We rely on users with CREATE EXTENSION privileges, usually administrators,
to manage the table and it's respective permission since in theory we could
have users execute any function in the table.