passcheck feature broken out into 5 commits (#5)
* Inital extension framework for uni_api

Breaking out the passcheck feature into individual commit to keep
it easier to track. This one just contains the initial extension
scaffolding

* Add gitignore, passcheck scaffolding

Included GUC to decide if passcheck feature should be enabled
(will be used in follow-up commits)

* Add SQL script for uni_api extension

Introduced enum for password_types, created a
table/general framework for features that can
be called out to SQL functions such as password_check.
Introduced a trigger to prevent function drops if the
object identity matches in table to prevent stale entries

* Introduced functionality for passcheck

This introduces the core-functionality for passcheck
and documents the behavior when certain things are missing
such as schema.

The overall idea is for the user to insert the function they
want to be associated with the passcheck feature. Then when
the hook is called, we will via SPI_exec read the table, and
call out to the function.

This requires SELECT privileges for all users on the bc.feature_info
table and EXECUTE permissions on the function inserted into the table.
We rely on users with CREATE EXTENSION privileges, usually administrators,
to manage the table and it's respective permission since in theory we could
have users execute any function in the table.

* Unit tests for passcheck feature

* Updated guc comment, removed unnecessary PG_MODUEL_MAGIC, ran pgindent

* Add licenses to all file, fixed passcheck.h, removed extra .out file