Force use of "pg_config --pgxs" for builds
Newer extensions tend to use this by default as it makes it
simpler for users to build extensions when the PostgreSQL
toolchain is installed. Yes, "USE_PGXS=1" is only one
additional item to add, but if one is trying to build
without docs, it is one less thing to forget.

Signed-off-by: Jonathan S. Katz <jkatz@amazon.com>