Apply project naming to pg_tle
This is an invasive patch to apply the project naming to pg_tle.
This updates the names of internal and external functions to match
the desired naming.

Note that only PostgreSQL system roles and schemas can be prefixed
with "pg_", so the administrative roles/schemas are called "pgtle".
This may or may not change in a future commit so capturing this
here.