Make all prefixes be "pgtle"
While it's OK to have the extension name as "pg_tle", we cannot
prefix schema/roles with "pg_" as that is reserved by PostgreSQL.
To make things consistent, anything that is user settable, e.g.
config parameters (GUCs), will now be "pgtle" to limit confusion.
We may also consider further refining everything to be "pgtle",
but that can be decided in a future commit.

fixes #47