Force all TLE extensions to set "trusted = false"
The word "trusted" for an extension in PostgreSQL has a specific
meaning, i.e. that an extension can escalate to a PostgreSQL
superuser role to perform actions. TLE extensions do not escalate
to superuser, but rather when CREATE EXTENSION is run, are
executed with the permissions of the user who runs the "CREATE"
command.

As such, we can delegate permission checks to regular PostgreSQL
access controls.

"trusted" in this context may mean something with TLE at a later
date.

Co-authored-by: Swaha Miller <swaha.miller@gmail.com>