Drop pgtle_staff role
The previous commit removes the need for a specalized role for
running CREATE / DROP EXTENSION commands on TLEs, as TLEs will
follow PostgreSQL's permission and execution model for
extensions. As such, this is just a superfluous permission that
complicates the overall management of TLEs.

A PostgreSQL user that runs "CREATE EXTENSION" is the owner
of that extensions. Other unprivileged users who do not own
that extension are unable to DROP it. Thus we do not need
to add an extra restriction with the pgtle_staff role.

Co-authored-by: Swaha Miller <swaha.miller@gmail.com>