Only allow alphanumeric characters and '_' separator in extension name
This fixes #98.

Since we rely on the executing the extension_name.control file function
to get the control contents, it makes it especially vulnerable to
additional injections based off of the extension name.

To prevent this restrict the allowed characters in extension name.
Theoretically we could make this more robust in terms of ControlFile
parsing the prosrc definition out, but let's revisit that in the future.