Gentler error handling when control file function has errors
This is likely due to someone trying to put in a contrived
comment or the like, but it can force undecipherable syntax errors.
We may need to do some additional work to sanitize inputs coming
from this function, but this at least gives the user guidance
on how to get out of trouble.

Also ensures we force the "directory" value to NULL, though we
do not reference it anywhere.

fixes #96