Force extensions to use server_side encoding
The "encoding" parameter in `install_extension` did not make
sense in this context, as encoding is either using client or
server side encoding. In a file-based extension, that can be
different.

As such, we remove the ability for the user to set the encoding
using the encoding parameter. A future commit may consider
checking if the user has explicitly set "client_encoding".