Fix segfault in password check hook
Not all compilers are smart enough to initialize a character
pointer to an empty string. Which is another way of saying that
this type was initialized improperly d37ddb1db and this now
handles it correctly.

This also hides some of the "magic" of setting the number of
arguments for a SPI_execute_with_args call by pulling it up
into the define level.