Ensure reload changes are propagated in tests
There is a known race condition in "pg_reload_conf()" where the
postmaster reload does not complete in time before the next command
is execute. Thus the current connection may be pulling server info
that is stale.

There are several solutions for this, but within the context of
the tests we can reconnect with "\c -", which is way better than
sleeping.

fixes #115