Ensure "sql_drop" trigger catches any reigstered function drop
Previously, this was only catching a direct "DROP FUNCTION" call
to a registered feature, instead of inspecting the object_type
to see what specific object was being dropped. For example, a
schema drop would also drop a function.

This now honors the behavior of the original call, though this
forces one to un-register a feature before calling "DROP EXTENSION".
Future work could make this work

fixes #130