Fix "requires" for file-based extensions
An issue with how file-based extension version files were copied
while building the dependency graph caused bogus names to be
available, and would prevent the proper requires tree from working
if there were both TLE / file-based extensions in the dependency
tree. This ensures the copy is handled in a way that does not
send garbage.

fixes #183