Better error handling when an extension doesn't exist (#192)
If the shared preload library for "pg_tle" is loaded but the
"pg_tle" extension is not activated and a user tries to install
an extension that does not exist, pg_tle would give an error
of the form "cannot find pgtle schema". However, while true, this
is not a helpful error message.

The issue was due to the ProcessUtility hook firing even though
the "pg_tle" extension itself was not active. To avoid this, we
add a check to the PU hook to only run its custom logic when
we know that "pg_tle" is installed.

fixes #187