Removed dead/unreachable return statements
This is in continuation of f8c7e95baf.

Code after ereport(ERROR) and elog(ERROR) calls is unreachable, so
remove such code.

In passing, add or remove blank lines, add curly braces, etc., to
improve readability. This is not a full sweep of the code base, just for
the parts that I was inspecting to look for dead code.