Disallow REPLACE or ALTER functions used or defined by pg_tle datatype APIs.
If a user-defined function is used by pgtle datatype API (i.e. create_base_type or create_operator_func), we will define a C-version function with the same name (different argument types).

Replacing or altering these functions can be error-prone. For example, if a user defined function is renamed, we won't be able to find it in the C version function because we rely on the same name as a mapping.

Dropping these functions is still allowed, and will drop the dependent objects cascade.