Disallow REPLACE or ALTER functions used or defined by pg_tle datatype APIs. (#219)
* Disallow REPLACE or ALTER functions used or defined by pg_tle datatype APIs.

If a user-defined function is used by pgtle datatype API (i.e. create_base_type or create_operator_func), we will define a C-version function with the same name (different argument types).

Replacing or altering these functions can be error-prone. For example, if a user defined function is renamed, we won't be able to find it in the C version function because we rely on the same name as a mapping.

Dropping these functions is still allowed, and will drop the dependent objects cascade.

* Improve error message for ALTER or REPLACE pgtle used functions.

1. Fix a bug where operator is not shown in the error message;
2. Add a few more test cases to verify that only pgtle used/defined operator functions (by create_operator_func) is restricted. If an operator function is used by `create_operator_func` API, it's not restricted from altering/replacing.

---------

Co-authored-by: Lyu Pan <lyup@amazon.com>