Fix the race condition that left an entry unprocessed
Previously the client backend checked num_requests to see if there are
available entries to load into. However, when a client finishes and
decrements num_requests, we don't know which index it was, so we can't
assume that the next index is free. This commit replaces num_requests
with an available_entry flag for each queue entry. The cliend backend
directly checks if the next entry is available instead.

This means that a client backend can be blocked by a single long-running
query if it happens to be occupying the next queue entry. We should
think about a way for client backends to "skip" entries if a later one
is available.