Fix stale entry bug
If a client backend terminates unexpectedly after setting
entry_available = false, it can block every subsequent client from using
that entry.

This commit adds a PID field to the request entry struct and allows
clients to check if the process currently holding the entry still
exists. If it doesn't, then it can set entry_available to true.

Signalling order is rearranged to prevent deadlocks where a backend is
terminated after entry_available is set to false but before signalling
anyone else. In addition to signalling the current client, background
workers signal the next waiting client to check if the current client
still exists.