Restructure pg_tle_create_base_type* C functions to fix a bug and improve readability.
When the dabase gets upgraded to a new version(solib is updated) but pg_tle version is not, the old 5-arg version SQL function is invoked by users, but the previous C function always 7 args.

In this chagne, a new pg_tle_create_base_type_7args C function is created and used to create the create_base_type SQL function.