Remove idx_offset optimization from clientauth. (#295)
There was a math bug with this optimization that caused incoming
connections to occassionally be missed by a worker.  This optimization
doesn't seem necessary and makes the code harder to understand and
maintain, so just get rid of it.

Update a TAP test for postgres 4f7f7b0 to pass CI on master.